SurveyResult를 result 디렉토리로 옮김
